WIP: cog-platform: moved cog-platform-fdo to cog-fdo-shell the module
* Moved platform/cog-platform-fdo.c to modules/cog-fdo-shell.c
* WL specific from platform/cog-platform-fdo.c moved to platform/pwl

TODO Refactoring of all the code labeled with TODO.
